test(storage): one tombstone-head builder for both seam suites
REVIEW ROUND 9. The duplication is this branch's own, created one round ago: the
same-sequence seam suite got its tombstone builder by copying the late-tombstone
suite's and adding two parameters.

WHY IT IS WORTH A COMMIT RATHER THAN A NOTE. Two copies of a fixture prove that
they agree, never that either one is correct, and a shared mistake passes both.
The rules this builder encodes are exactly the ones a head-codec change moves:
which active-only fields must be deleted, which projection counters must be
zeroed, and that the result has to be re-validated rather than trusted. A drift
would not have failed loudly -- one suite would have kept building a shape the
codec no longer accepts, and the codec's refusal is phrased like a domain refusal
rather than a fixture one. Two seats on this branch have already lost a wall to
that confusion.

`buildTombstoneHeadFromPredecessorV1` takes the predecessor, an optional version
and optional field overrides. Both suites call it; five imports the extraction
orphaned are removed.

TWO THINGS DELIBERATELY NOT MERGED INTO IT.

The version-is-a-parameter reasoning travels WITH the builder, because it is a
property of the builder rather than of either caller: the binding predicate wants
a version strictly GREATER than the predecessor's, not adjacent, so hard-coding
`predecessor.version + 1` reaches exactly one version relation and makes the
other two look unconstructible. That cost a wall before it was parameterized.

The load-bearing coincidence stays in the same-sequence suite, because it is a
fact about that suite's chain and not a general guarantee: the closure resolves
the predecessor through the tombstone's own `previousHeadDigest` against the
ancestry while the mint graph supplies only the owned-subject table, and those
agree only because the current head is a spread of the base. Promoting it into
the helper would have made a coincidence look like a contract.

Both seam suites green at 23 tests; the four verdict-diff lanes that pin lines
into these files still resolve, re-checked by the discovery tool rather than by
assuming a test-only change cannot move a citation.
